Removed timeout parameter in checks func

+++ database.sql
@@ -7,7 +7,7 @@ CREATE TABLE IF NOT EXISTS urls (
 CREATE TABLE IF NOT EXISTS url_checks (
     id BIGINT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
     url_id BIGINT REFERENCES urls (id),
-    status_code INT,
+    status_code SMALLINT,
     h1 VARCHAR(255),
     title VARCHAR(255),
     description TEXT,
